changed header structure, improved regex comparison
