+++ include/comparator/equality.hpp
@@ -19,15 +19,15 @@
  }
  */
 
-#ifndef SCTF_COMPARATOR_EQUALS_HPP
-#define SCTF_COMPARATOR_EQUALS_HPP
+#ifndef SCTF_COMPARATOR_EQUALITY_HPP
+#define SCTF_COMPARATOR_EQUALITY_HPP
 
 #include <algorithm>
 #include <cmath>
 #include <limits>
 
 #include "common/traits.hpp"
-#include "comparator/comparators.hpp"
+#include "comparator/comparator.hpp"
 
 namespace sctf
 {
@@ -77,10 +77,57 @@ public:
                               to_string(expected_value));
     }
 };
+
+class unequals
+{
+    static constexpr char const* m_cmp_str     = "to be unequals";
+    static constexpr char const* m_neg_cmp_str = "to be not unequals";
+    bool                         m_neg         = false;
+
+public:
+    unequals& operator!() {
+        m_neg = !m_neg;
+        return *this;
+    }
+
+    template<typename V, typename E = V, ENABLE_IF(NOT IS_FLOAT(V) AND IS_UNEQUAL_COMPARABLE(V, E))>
+    comparison operator()(V const& actual_value, E const& expected_value) {
+        return (actual_value != expected_value) != m_neg ?
+                   SUCCESS :
+                   comparison(m_neg ? m_neg_cmp_str : m_cmp_str, to_string(actual_value),
+                              to_string(expected_value));
+    }
+
+    template<typename V, typename E = V,
+             ENABLE_IF(NOT IS_FLOAT(V) AND NOT IS_UNEQUAL_COMPARABLE(V, E)
+                           AND                 IS_EQUAL_COMPARABLE(V, E))>
+    comparison operator()(V const& actual_value, E const& expected_value) {
+        return (actual_value == expected_value) != m_neg ?
+                   comparison(m_neg ? m_neg_cmp_str : m_cmp_str, to_string(actual_value),
+                              to_string(expected_value)) :
+                   SUCCESS;
+    }
+
+    template<typename V, typename E = V, ENABLE_IF(IS_FLOAT(V) AND IS_FLOAT(E))>
+    comparison operator()(V const& actual_value, E const& expected_value) {
+#if defined(SCTF_EXTERN_EPSILON) || defined(SCTF_EPSILON)
+        static V epsilon_ = static_cast<V>(epsilon);
+#else
+        static V epsilon_ = std::numeric_limits<V>::epsilon();
+#endif
+        return (std::abs(actual_value - expected_value) <=
+                std::max(std::abs(actual_value), std::abs(expected_value)) * epsilon_) != m_neg ?
+                   comparison(m_neg ? m_neg_cmp_str : m_cmp_str, to_string(actual_value),
+                              to_string(expected_value)) :
+                   SUCCESS;
+    }
+};
 }  // namespace private_
 }  // namespace sctf
 
+PROVIDE_COMPARATOR(unequals, UNEQUALS)
+PROVIDE_COMPARATOR(unequals, NE)
 PROVIDE_COMPARATOR(equals, EQUALS)
 PROVIDE_COMPARATOR(equals, EQ)
 
-#endif  // SCTF_COMPARATOR_EQUALS_HPP
+#endif  // SCTF_COMPARATOR_EQUALITY_HPP
