+++ include/comparator/ordering.hpp
@@ -19,14 +19,19 @@
  }
  */
 
-#ifndef SCTF_COMPARATOR_LESS_HPP
-#define SCTF_COMPARATOR_LESS_HPP
+#ifndef SCTF_COMPARATOR_ORDERING_HPP
+#define SCTF_COMPARATOR_ORDERING_HPP
 
-#include "comparator/comparators.hpp"
+#include "comparator/comparator.hpp"
+
+COMPARATOR(greater_than, "greater than", actual_value > expected_value)
+PROVIDE_COMPARATOR(greater_than, GREATER_THAN)
+PROVIDE_COMPARATOR(greater_than, GREATER)
+PROVIDE_COMPARATOR(greater_than, GT)
 
 COMPARATOR(less_than, "less than", actual_value < expected_value)
 PROVIDE_COMPARATOR(less_than, LESS_THAN)
 PROVIDE_COMPARATOR(less_than, LESS)
 PROVIDE_COMPARATOR(less_than, LT)
 
-#endif  // SCTF_COMPARATOR_LESS_HPP
+#endif  // SCTF_COMPARATOR_ORDERING_HPP
