+++ include/sctf.hpp
@@ -25,13 +25,10 @@
 #define SCFT_VERSION "1.2-rc2"
 
 #include "common/types.hpp"
-#include "comparator/equals.hpp"
-#include "comparator/greater.hpp"
-#include "comparator/inrange.hpp"
-#include "comparator/less.hpp"
-#include "comparator/like.hpp"
-#include "comparator/match.hpp"
-#include "comparator/unequals.hpp"
+#include "comparator/equality.hpp"
+#include "comparator/ordering.hpp"
+#include "comparator/range.hpp"
+#include "comparator/regex.hpp"
 #include "reporter/console_reporter.hpp"
 #include "reporter/markdown_reporter.hpp"
 #include "reporter/xml_reporter.hpp"
